Initial cut at enabling multi level projects to work with BNDTools
Updated a new project's build.xml file so it works with a non-flat
directory structure for bundle projects.
Added processing to generate the right text into the build.xml file
for a new project.

When creating a new project, getProjectLocationURI() needed to return
null for projects at the top level, but if deeper in the workspace path,
it needed to provide the absolute path to the project. Seemed like strange
eclipse-ism stuff.

Additionally, pathing adjustments to use BND's workspace to find projects.

Probably need to understand what the exact reason is for using
eclipse pathing in all of these locations. Perhaps I'm missing
something important and shouldn't be making these sorts of
changes.